refactor(industrial): Stage 1 — extract kirra-industrial crate
ADR-0035 Stage 1. With the Stage 0 foundation in place (OperationalCommand +
the action_claim cluster now in kirra-policy-types), the industrial layer is a
clean leaf, so move it out of the heavy root crate.

Moves VERBATIM into the new lean kirra-industrial crate (behaviour-identical
pure moves — the only edits are the three cross-crate import rewrites):
- src/protocol_adapter.rs → kirra-industrial (IndustrialProtocol / IndustrialEvent,
  the unified + per-protocol event→claim mapping and posture evaluation, freshness
  classification, the QW2 replay-key slug)
- src/adapters/{mod,canopen,dnp3,ethernet_ip}.rs → kirra-industrial::adapters
  (the IndustrialAdapter trait, ScalarType/BoundSpec, and the CANopen / DNP3 /
  EtherNet-IP fieldbus decoders + their env-configured magnitude bounds)

Import rewrites (the whole cross-crate surface):
- crate::gateway::policy::OperationalCommand → kirra_policy_types::OperationalCommand
- crate::action_filter::{evaluate_action_claim,ActionClaim,ActionDecision}
    → kirra_policy_types::action_claim::{..}
- crate::verifier::FleetPosture → kirra_core::FleetPosture

Confirmed those are the ONLY non-internal references — every other path is the
crate's own adapters::* — so kirra-industrial depends only on kirra-policy-types
+ kirra-core (+ serde/serde_json/tracing). A clean leaf, no service/DB deps.

The root crate keeps src/protocol_adapter.rs and src/adapters.rs as pub-use
re-export shims (whole-module globs, so the child-module paths
kirra_verifier::adapters::{canopen,dnp3,ethernet_ip}::* and the init_*_from_env
entry points all resolve unchanged). The service bin's rich consumption surface
is untouched.

Verified: kirra-industrial 105 tests green, root package test suite green
(incl. the bin's 5 industrial + 8 dnp3-audit/replay handler tests exercising the
shimmed paths), both trees fmt-clean, quality-guardrails satisfied, orphan-core
green. All four lockfiles (root + detached fuzz/parko/kirra-verifier-pg) updated
minimally for the new crate — clean additions, no churn.
