+++ crates/kirra-industrial/src/adapters/dnp3.rs
@@ -2,7 +2,7 @@
 
 use std::sync::OnceLock;
 
-use crate::gateway::policy::OperationalCommand;
+use kirra_policy_types::OperationalCommand;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
