+++ crates/kirra-industrial/src/adapters/ethernet_ip.rs
@@ -1,6 +1,6 @@
 // src/adapters/ethernet_ip.rs
 
-use crate::gateway::policy::OperationalCommand;
+use kirra_policy_types::OperationalCommand;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
